Begun Pokédex View, more details below...
The PokedexViewState class derives from AutomaticKeepAliveClientMixin<PokedexView>
in order to keep 'alive' even if it is not the current state.
The basic JSON files from the Pokemons are loaded at the beginning of
the app.
The 'Tools' class provides some static functions useful for development
(only for Strings for the moment).

The app loads the 20 first Pokemon of the National Pokédex for the moment.